feat(db): P3 pre-fix — silver partials real->float8 (exact RAW-equivalent)
The silver.equipment_metrics_* partials were `real` (inherited from the float4
source cols), losing ~4.8e-6 relative precision on large buckets and compounding
through the telescoping rollup. Redefine as float8 (cast source ::float8 in tier-1)
so silver is EXACTLY equal to a float8 direct-RAW sum BEFORE any aggregate serving
fn repoints onto it in P3.

- 10_p3pre_silver_float8_partials.sql: drop+recreate the 4-tier family (float8),
  drop/recreate serving.machine_speed (tier-1 dependent) around it, re-add policies.
- 04_p1: greenfield DDL now casts ::float8 too (fresh apply is correct).
- 05_p1: fix the cosmetic 4/4 policy assertion join (jobs.hypertable_name = user view).

HP-1 re-proven on staging (frozen window 07-08..09-07, full dim grouping):
  1min vs direct-RAW float8 -> 1,388,365 rows, 0 presence/count/sum mismatch, max abs diff = 0.
  telescoping 1min->1hour -> 0 count mismatch, max abs 7.3e-12 (float8 assoc noise).

+++ db/migrations/analytics-clean-schema/04_p1_silver_metrics_family.sql
@@ -8,9 +8,12 @@ CREATE MATERIALIZED VIEW IF NOT EXISTS silver.equipment_metrics_1min
 WITH (timescaledb.continuous, timescaledb.materialized_only = true) AS
 SELECT time_bucket('1 minute', ts_value) AS bucket,
   id_equipment, id_enterprise, id_site, id_area, tp_equipment,
-  sum(net_production_incr) AS sum_net, sum(gross_production_incr) AS sum_gross,
-  sum(scrap_incr) AS sum_scrap, sum(speed) AS sum_speed, count(speed) AS cnt_speed,
-  count(*) AS cnt_rows, max(speed) AS max_speed, max(ideal_production_speed) AS ideal_production_speed
+  sum(net_production_incr::float8) AS sum_net, sum(gross_production_incr::float8) AS sum_gross,
+  sum(scrap_incr::float8) AS sum_scrap, sum(speed::float8) AS sum_speed, count(speed) AS cnt_speed,
+  count(*) AS cnt_rows, max(speed::float8) AS max_speed, max(ideal_production_speed) AS ideal_production_speed
+  -- NOTE: partials are float8 (not the float4 source type) so the telescoping rollup is exactly
+  -- equal to a float8 direct-RAW sum. float4 partials lost ~4.8e-6 rel precision on large buckets
+  -- (fixed 2026-09-08, P3 pre-fix — see 10_p3pre_silver_float8_partials.sql).
 FROM public.equipment_values WHERE tp_equipment IS NOT NULL
 GROUP BY 1, id_equipment, id_enterprise, id_site, id_area, tp_equipment
 WITH NO DATA;
